refactor(check): move check component from molecules to atoms
ref: #36
