+++ index.js
@@ -13,7 +13,7 @@ export { default as AvatarHeader } from './molecules/AvatarHeader'
 export { default as Toolbar } from './molecules/Toolbar'
 export { default as Task } from './molecules/Task'
 export { default as Buttons } from './molecules/Buttons'
-export { default as Check } from './molecules/Check'
+export { default as Check } from './atoms/Check'
 export { default as SeparatorLine } from './atoms/SeparatorLine'
 export { default as TasksCounter } from './molecules/TasksCounter'
 export { default as Toggel } from './atoms/Toggel'
