+++ molecules/Task/Task.stories.js
@@ -11,20 +11,22 @@ import {
 const Template = getTemplate(Task, styles)
 const ListTemplate = getListTemplate(Task, styles)
 
+const args = {
+  children: 'Sir Goblin-town feels characters ruined precautions.',
+  onCheck: () => {},
+  onDelete: () => {},
+  defaultChecked: false,
+}
+
 export default {
   title: 'Molecules/Task',
   component: Task,
-  args: {
-    children: 'Sir Goblin-town feels characters ruined precautions.',
-  },
+  args,
   argTypes: {
     type: getOptionsArgTypes(options.types),
     onCheck: { defaultValue: null },
     onDelete: { defaultValue: null },
   },
-}
-const args = {
-  children: 'Sir Goblin-town feels characters ruined precautions.',
 }
 
 export const Default = Template.bind({})
@@ -32,15 +34,16 @@ Default.args = { ...args, id: 'default' }
 
 export const Long = Template.bind({})
 Long.args = {
+  ...args,
   id: 'long',
   children:
     "Swords are no more use here. Moments shiny rings Cirith nowhere garrison ax pointy endless mention trousers. Retribution Frodo's wrong? Them leaf question require spared houses sapphire Balin. Kind pouring Lothlórien ragged retribution covers lair less goings recognizes unleash. Whence Precious token hidden hire bandy make farming. Shards farmer unfriendly handle Bain gongs. Mend recall sorry oncoming roll unmade cries visitors Lobelia Sackville-Baggins was bidden. Dank rather between short doomed Mellon tomb consequence girl crescent attacking.",
 }
 
-export const Checked = Template.bind({})
-Checked.args = {
+export const CheckedTasks = Template.bind({})
+CheckedTasks.args = {
   ...args,
-  defaultIsChecked: true,
+  defaultChecked: true,
   id: `checked-true`,
 }
